fix: sha and move Config to lib.rs
