+++ extensions/native/circuit/src/extension.rs
@@ -9,16 +9,12 @@ use native_vectorized_adapter::{NativeVectorizedAdapterAir, NativeVectorizedAdap
 use openvm_circuit::{
     arch::{
         AirInventory, AirInventoryError, ChipInventory, ChipInventoryError, ExecutionBridge,
-        ExecutorInventoryBuilder, ExecutorInventoryError, InitFileGenerator, MatrixRecordArena,
-        MemoryConfig, RowMajorMatrixArena, SystemConfig, VmBuilder, VmChipComplex,
-        VmCircuitExtension, VmExecutionExtension, VmProverExtension,
+        ExecutorInventoryBuilder, ExecutorInventoryError, RowMajorMatrixArena, VmCircuitExtension,
+        VmExecutionExtension, VmProverExtension,
     },
-    system::{
-        memory::SharedMemoryHelper, SystemChipInventory, SystemCpuBuilder, SystemExecutor,
-        SystemPort,
-    },
+    system::{memory::SharedMemoryHelper, SystemPort},
 };
-use openvm_circuit_derive::{AnyEnum, InsExecutorE1, InsExecutorE2, InstructionExecutor, VmConfig};
+use openvm_circuit_derive::{AnyEnum, InsExecutorE1, InsExecutorE2, InstructionExecutor};
 use openvm_instructions::{program::DEFAULT_PC_STEP, LocalOpcode, PhantomDiscriminant};
 use openvm_native_compiler::{
     CastfOpcode, FieldArithmeticOpcode, FieldExtensionOpcode, FriOpcode, NativeBranchEqualOpcode,
@@ -26,10 +22,7 @@ use openvm_native_compiler::{
     NativeRangeCheckOpcode, Poseidon2Opcode, VerifyBatchOpcode, BLOCK_LOAD_STORE_SIZE,
 };
 use openvm_poseidon2_air::Poseidon2Config;
-use openvm_rv32im_circuit::{
-    BranchEqualCoreAir, Rv32I, Rv32IExecutor, Rv32ImCpuProverExt, Rv32Io, Rv32IoExecutor, Rv32M,
-    Rv32MExecutor,
-};
+use openvm_rv32im_circuit::BranchEqualCoreAir;
 use openvm_stark_backend::{
     config::{StarkGenericConfig, Val},
     p3_field::{Field, PrimeField32},
@@ -46,71 +39,6 @@ use crate::{
     phantom::*,
     *,
 };
-
-#[derive(Clone, Debug, derive_new::new, VmConfig, Serialize, Deserialize)]
-pub struct NativeConfig {
-    #[config(executor = "SystemExecutor<F>")]
-    pub system: SystemConfig,
-    #[extension(generics = true)]
-    pub native: Native,
-}
-
-impl NativeConfig {
-    pub fn aggregation(num_public_values: usize, max_constraint_degree: usize) -> Self {
-        Self {
-            system: SystemConfig::new(
-                max_constraint_degree,
-                MemoryConfig::aggregation(),
-                num_public_values,
-            )
-            .with_max_segment_len((1 << 24) - 100),
-            native: Default::default(),
-        }
-    }
-}
-
-// Default implementation uses no init file
-impl InitFileGenerator for NativeConfig {}
-
-pub struct NativeCpuBuilder(pub NativeConfig);
-impl<E, SC> VmBuilder<E> for NativeCpuBuilder
-where
-    SC: StarkGenericConfig,
-    E: StarkEngine<SC = SC, PB = CpuBackend<SC>, PD = CpuDevice<SC>>,
-    Val<SC>: PrimeField32,
-{
-    type VmConfig = NativeConfig;
-    type SystemChipInventory = SystemChipInventory<SC>;
-    type RecordArena = MatrixRecordArena<Val<SC>>;
-
-    fn config(&self) -> &Self::VmConfig {
-        &self.0
-    }
-
-    fn create_chip_complex(
-        &self,
-        circuit: AirInventory<SC>,
-    ) -> Result<
-        VmChipComplex<SC, Self::RecordArena, E::PB, Self::SystemChipInventory>,
-        ChipInventoryError,
-    > {
-        let config = &self.0;
-        let system = SystemCpuBuilder(config.system.clone());
-        let mut chip_complex = VmBuilder::<E>::create_chip_complex(&system, circuit)?;
-        let inventory = &mut chip_complex.inventory;
-        VmProverExtension::<E, _, _>::extend_prover(
-            &NativeCpuProverExt,
-            &config.native,
-            inventory,
-        )?;
-        Ok(chip_complex)
-    }
-}
-impl From<NativeCpuBuilder> for NativeConfig {
-    fn from(builder: NativeCpuBuilder) -> Self {
-        builder.0
-    }
-}
 
 // ============ VmExtension Implementations ============
 
@@ -580,38 +508,6 @@ where
     }
 }
 
-#[derive(Clone, Debug, VmConfig, derive_new::new, Serialize, Deserialize)]
-pub struct Rv32WithKernelsConfig {
-    #[config(executor = "SystemExecutor<F>")]
-    pub system: SystemConfig,
-    #[extension]
-    pub rv32i: Rv32I,
-    #[extension]
-    pub rv32m: Rv32M,
-    #[extension]
-    pub io: Rv32Io,
-    #[extension(generics = true)]
-    pub native: Native,
-    #[extension]
-    pub castf: CastFExtension,
-}
-
-impl Default for Rv32WithKernelsConfig {
-    fn default() -> Self {
-        Self {
-            system: SystemConfig::default().with_continuations(),
-            rv32i: Rv32I,
-            rv32m: Rv32M::default(),
-            io: Rv32Io,
-            native: Native,
-            castf: CastFExtension,
-        }
-    }
-}
-
-// Default implementation uses no init file
-impl InitFileGenerator for Rv32WithKernelsConfig {}
-
 // Pre-computed maximum trace heights for NativeConfig. Found by doubling
 // the actual trace heights of kitchen-sink leaf verification (except for
 // VariableRangeChecker, which has a fixed height).
@@ -619,49 +515,3 @@ pub const NATIVE_MAX_TRACE_HEIGHTS: &[u32] = &[
     4194304, 4, 128, 2097152, 8388608, 4194304, 262144, 2097152, 16777216, 2097152, 8388608,
     262144, 2097152, 1048576, 4194304, 65536, 262144,
 ];
-
-pub struct Rv32WithKernelsCpuBuilder(pub Rv32WithKernelsConfig);
-
-impl<E, SC> VmBuilder<E> for Rv32WithKernelsCpuBuilder
-where
-    SC: StarkGenericConfig,
-    E: StarkEngine<SC = SC, PB = CpuBackend<SC>, PD = CpuDevice<SC>>,
-    Val<SC>: PrimeField32,
-{
-    type VmConfig = Rv32WithKernelsConfig;
-    type SystemChipInventory = SystemChipInventory<SC>;
-    type RecordArena = MatrixRecordArena<Val<SC>>;
-
-    fn config(&self) -> &Self::VmConfig {
-        &self.0
-    }
-
-    fn create_chip_complex(
-        &self,
-        circuit: AirInventory<SC>,
-    ) -> Result<
-        VmChipComplex<SC, Self::RecordArena, E::PB, Self::SystemChipInventory>,
-        ChipInventoryError,
-    > {
-        let config = &self.0;
-        let system = SystemCpuBuilder(config.system.clone());
-        let mut chip_complex = VmBuilder::<E>::create_chip_complex(&system, circuit)?;
-        let inventory = &mut chip_complex.inventory;
-        VmProverExtension::<E, _, _>::extend_prover(&Rv32ImCpuProverExt, &config.rv32i, inventory)?;
-        VmProverExtension::<E, _, _>::extend_prover(&Rv32ImCpuProverExt, &config.rv32m, inventory)?;
-        VmProverExtension::<E, _, _>::extend_prover(&Rv32ImCpuProverExt, &config.io, inventory)?;
-        VmProverExtension::<E, _, _>::extend_prover(
-            &NativeCpuProverExt,
-            &config.native,
-            inventory,
-        )?;
-        VmProverExtension::<E, _, _>::extend_prover(&NativeCpuProverExt, &config.castf, inventory)?;
-        Ok(chip_complex)
-    }
-}
-
-impl From<Rv32WithKernelsCpuBuilder> for Rv32WithKernelsConfig {
-    fn from(builder: Rv32WithKernelsCpuBuilder) -> Self {
-        builder.0
-    }
-}
