+++ extensions/sha256/circuit/src/lib.rs
@@ -1,5 +1,96 @@
+use std::result::Result;
+
+use openvm_circuit::{
+    arch::{
+        AirInventory, ChipInventoryError, InitFileGenerator, MatrixRecordArena, SystemConfig,
+        VmBuilder, VmChipComplex, VmProverExtension,
+    },
+    system::{SystemChipInventory, SystemCpuBuilder, SystemExecutor},
+};
+use openvm_circuit_derive::VmConfig;
+use openvm_rv32im_circuit::{
+    Rv32I, Rv32IExecutor, Rv32ImCpuProverExt, Rv32Io, Rv32IoExecutor, Rv32M, Rv32MExecutor,
+};
+use openvm_stark_backend::{
+    config::{StarkGenericConfig, Val},
+    p3_field::PrimeField32,
+    prover::cpu::{CpuBackend, CpuDevice},
+};
+use openvm_stark_sdk::engine::StarkEngine;
+use serde::{Deserialize, Serialize};
+
 mod sha256_chip;
 pub use sha256_chip::*;
 
 mod extension;
 pub use extension::*;
+
+#[derive(Clone, Debug, VmConfig, derive_new::new, Serialize, Deserialize)]
+pub struct Sha256Rv32Config {
+    #[config(executor = "SystemExecutor<F>")]
+    pub system: SystemConfig,
+    #[extension]
+    pub rv32i: Rv32I,
+    #[extension]
+    pub rv32m: Rv32M,
+    #[extension]
+    pub io: Rv32Io,
+    #[extension]
+    pub sha256: Sha256,
+}
+
+impl Default for Sha256Rv32Config {
+    fn default() -> Self {
+        Self {
+            system: SystemConfig::default().with_continuations(),
+            rv32i: Rv32I,
+            rv32m: Rv32M::default(),
+            io: Rv32Io,
+            sha256: Sha256,
+        }
+    }
+}
+
+// Default implementation uses no init file
+impl InitFileGenerator for Sha256Rv32Config {}
+
+pub struct Sha256Rv32CpuBuilder(pub Sha256Rv32Config);
+
+impl<E, SC> VmBuilder<E> for Sha256Rv32CpuBuilder
+where
+    SC: StarkGenericConfig,
+    E: StarkEngine<SC = SC, PB = CpuBackend<SC>, PD = CpuDevice<SC>>,
+    Val<SC>: PrimeField32,
+{
+    type VmConfig = Sha256Rv32Config;
+    type SystemChipInventory = SystemChipInventory<SC>;
+    type RecordArena = MatrixRecordArena<Val<SC>>;
+
+    fn config(&self) -> &Self::VmConfig {
+        &self.0
+    }
+
+    fn create_chip_complex(
+        &self,
+        circuit: AirInventory<SC>,
+    ) -> Result<
+        VmChipComplex<SC, Self::RecordArena, E::PB, Self::SystemChipInventory>,
+        ChipInventoryError,
+    > {
+        let config = &self.0;
+        let system = SystemCpuBuilder(config.system.clone());
+        let mut chip_complex = VmBuilder::<E>::create_chip_complex(&system, circuit)?;
+        let inventory = &mut chip_complex.inventory;
+        VmProverExtension::<E, _, _>::extend_prover(&Rv32ImCpuProverExt, &config.rv32i, inventory)?;
+        VmProverExtension::<E, _, _>::extend_prover(&Rv32ImCpuProverExt, &config.rv32m, inventory)?;
+        VmProverExtension::<E, _, _>::extend_prover(&Rv32ImCpuProverExt, &config.io, inventory)?;
+        VmProverExtension::<E, _, _>::extend_prover(&Sha2CpuProverExt, &config.sha256, inventory)?;
+        Ok(chip_complex)
+    }
+}
+
+impl From<Sha256Rv32CpuBuilder> for Sha256Rv32Config {
+    fn from(builder: Sha256Rv32CpuBuilder) -> Self {
+        builder.0
+    }
+}
